requirements.txt
flask
openai
python-dotenv
notion-client
requests
gunicorn